# Performance updates to runs helpers