Now preferring branchless code in a few places, which also allows us to avoid a `memset()` call in `vec_detect_runs()` and optimize memory in `vec_locate_runs()`

Seems to help a lot in the case where we have large amount of very short runs (so the branches would be fairly unpredictable). This is the first example below. Doesn't seem to ever really hurt, which is promising, and doesn't make the code any less clear.

``` r
set.seed(123)

# large number of short runs
x <- sample(2, 1e7, replace = TRUE)

bench::mark(
  vec_identify_runs(x),
  vec_detect_runs(x, start = TRUE),
  vec_detect_runs(x, start = FALSE),
  vec_locate_runs(x, start = TRUE),
  vec_locate_runs(x, start = FALSE),
  iterations = 100,
  check = FALSE
)

# main
#> # A tibble: 5 × 6
#>   expression                            min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                8.9ms  9.58ms     102.     38.1MB     43.7
#> 2 vec_detect_runs(x, start = TRUE)   41.2ms 43.29ms      23.0    76.3MB     23.0
#> 3 vec_detect_runs(x, start = FALSE)  44.7ms 46.39ms      21.3    76.3MB     21.3
#> 4 vec_locate_runs(x, start = TRUE)   38.6ms 39.98ms      24.9    57.2MB     12.8
#> 5 vec_locate_runs(x, start = FALSE)  38.1ms 39.38ms      25.3    57.2MB     12.4

# This PR
#> # A tibble: 5 × 6
#>   expression                            min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)               9.07ms  9.87ms     101.     38.1MB     52.0
#> 2 vec_detect_runs(x, start = TRUE)  15.47ms 16.87ms      57.3    76.3MB     79.1
#> 3 vec_detect_runs(x, start = FALSE)  15.5ms 16.11ms      61.4    76.3MB     61.4
#> 4 vec_locate_runs(x, start = TRUE)  15.23ms 16.11ms      61.7    38.1MB     20.6
#> 5 vec_locate_runs(x, start = FALSE) 15.45ms 16.69ms      60.0    38.1MB     20.0

# sorting means we only have 2 very long runs
x <- sort(x)

bench::mark(
  vec_identify_runs(x),
  vec_detect_runs(x, start = TRUE),
  vec_detect_runs(x, start = FALSE),
  vec_locate_runs(x, start = TRUE),
  vec_locate_runs(x, start = FALSE),
  iterations = 100,
  check = FALSE
)

# main
#> # A tibble: 5 × 6
#>   expression                            min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)               8.57ms  9.01ms     110.     38.1MB     36.6
#> 2 vec_detect_runs(x, start = TRUE)  17.57ms 18.91ms      51.6    76.3MB     84.1
#> 3 vec_detect_runs(x, start = FALSE) 19.84ms 20.63ms      47.2    76.3MB     91.6
#> 4 vec_locate_runs(x, start = TRUE)  13.65ms 14.19ms      69.7    38.1MB     17.4
#> 5 vec_locate_runs(x, start = FALSE) 13.57ms 14.37ms      68.8    38.1MB     17.2

# This PR
#> # A tibble: 5 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                8.72ms  9.3ms     107.     38.1MB     57.5
#> 2 vec_detect_runs(x, start = TRUE)   14.58ms 16.3ms      61.2    76.3MB     63.7
#> 3 vec_detect_runs(x, start = FALSE)  15.52ms 15.9ms      62.0    76.3MB     62.0
#> 4 vec_locate_runs(x, start = TRUE)   14.96ms   16ms      62.2    38.1MB     20.7
#> 5 vec_locate_runs(x, start = FALSE)  15.16ms 16.2ms      61.6    38.1MB     20.5

# - around 12 mil values
# - 500k short runs with size from 1-50
x <- vec_rep_each(1:500000, sample(1:50, 500000, replace = TRUE))
length(x)
#> [1] 12755775

bench::mark(
  vec_identify_runs(x),
  vec_detect_runs(x, start = TRUE),
  vec_detect_runs(x, start = FALSE),
  vec_locate_runs(x, start = TRUE),
  vec_locate_runs(x, start = FALSE),
  iterations = 100,
  check = FALSE
)

# main
#> # A tibble: 5 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                11.1ms 12.1ms      82.3    48.7MB     82.3
#> 2 vec_detect_runs(x, start = TRUE)    22.6ms 28.8ms      33.8    97.3MB     36.6
#> 3 vec_detect_runs(x, start = FALSE)   28.5ms 31.2ms      32.3    97.3MB     32.3
#> 4 vec_locate_runs(x, start = TRUE)    21.6ms 23.1ms      42.8    50.6MB     22.0
#> 5 vec_locate_runs(x, start = FALSE)   21.2ms 22.2ms      44.6    50.6MB     22.0

# This PR
#> # A tibble: 5 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                10.9ms 11.9ms      83.4    48.7MB     29.3
#> 2 vec_detect_runs(x, start = TRUE)    19.7ms 21.8ms      45.7    97.3MB     58.1
#> 3 vec_detect_runs(x, start = FALSE)   20.1ms 20.9ms      47.4    97.3MB     84.2
#> 4 vec_locate_runs(x, start = TRUE)    19.2ms 20.3ms      48.8    48.7MB     14.6
#> 5 vec_locate_runs(x, start = FALSE)   18.8ms 20.1ms      49.3    48.7MB     12.3

# - around 15 mil values
# - 20k medium sized runs with size from 500-1000
x <- vec_rep_each(1:20000, sample(500:1000, 20000, replace = TRUE))
length(x)
#> [1] 15033098

bench::mark(
  vec_identify_runs(x),
  vec_detect_runs(x, start = TRUE),
  vec_detect_runs(x, start = FALSE),
  vec_locate_runs(x, start = TRUE),
  vec_locate_runs(x, start = FALSE),
  iterations = 100,
  check = FALSE
)

# main
#> # A tibble: 5 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                12.9ms 13.5ms      73.4    57.3MB     36.2
#> 2 vec_detect_runs(x, start = TRUE)    25.9ms 27.1ms      35.6   114.7MB     43.6
#> 3 vec_detect_runs(x, start = FALSE)   28.9ms 29.8ms      32.6   114.7MB     32.6
#> 4 vec_locate_runs(x, start = TRUE)    20.7ms 21.6ms      46.0    57.4MB     15.3
#> 5 vec_locate_runs(x, start = FALSE)   20.4ms 21.3ms      46.7    57.4MB     15.6

# This PR
#> # A tibble: 5 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                        <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)                  13ms 13.9ms      70.4    57.3MB     37.9
#> 2 vec_detect_runs(x, start = TRUE)    22.8ms 24.4ms      38.6   114.7MB     40.2
#> 3 vec_detect_runs(x, start = FALSE)     23ms 25.1ms      39.0   114.7MB     39.0
#> 4 vec_locate_runs(x, start = TRUE)    23.4ms 26.5ms      36.0    57.3MB     12.0
#> 5 vec_locate_runs(x, start = FALSE)     21ms 23.6ms      42.2    57.3MB     14.1
```